requirements.txt
sqlalchemy==2.0.35

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

uvicorn
fastapi
pydantic

pytz
python-dateutil